+++ game/Scripts/Math/MathScore.gd
@@ -1,7 +1,10 @@
 extends Control
 
+var dataMgr
 var logMgr
 var panelMgr
+
+onready var ppAverage: Label = $VBoxContainer/Menu0/LabelAverage
 
 var arg_dic: Dictionary
 
@@ -10,6 +13,7 @@ export (PackedScene) var tips_prefab
 func _ready():
 	panelMgr = get_node("/root/PanelMgr")
 	logMgr = get_node("/root/GLog")
+	dataMgr = get_node("/root/GData")
 
 
 func _setRectSize(ssize: Vector2):
@@ -25,8 +29,16 @@ func _on_BtnBack_button_down():
 
 func _panel_set_dic(dic: Dictionary):
 	arg_dic = dic
-	
+	var score_sum: float = 0
+	var score_num: int = 0
 	for i in arg_dic:
 		var node = tips_prefab.instance()
 		$VBoxContainer/ScoreList.add_child(node)
 		node._set_data(arg_dic[i])
+		if arg_dic[i].size() == 4:
+			score_sum += arg_dic[i][3]
+			score_num += 1
+	
+	ppAverage.text = "%.2f" % [score_sum / score_num] + "s"
+	dataMgr._save()
+